Make use of teleporting to allow all modal components to be stored together
