+++ src/components/Teams/TeamHeader.vue
@@ -1,5 +1,6 @@
 <script setup>
   import { useTeamStore } from "@/stores/TeamStore";
+  import AddMemberModal from "@/components/Teams/AddMemberModal.vue";
 
   let team = useTeamStore();
 </script>
@@ -7,13 +8,7 @@
 <template>
     <header class="flex justify-between">
         <div>
-        <button
-            class="bg-green-500 hover:bg-green-600 text-white px-4 py-2 rounded disabled:bg-gray-400"
-            :disabled="! team.spotsRemaining"
-            @click="$emit('add')"
-        >
-            Add Member ({{ team.spotsRemaining }} Spots Left)
-        </button>
+            <AddMemberModal />
         </div>
 
         <div>
